Bank Account
- Accounts being written
- account information is going well

Fixes Needed
- When writting to the file it gets overwritten

+++ src/CreateAccount.java
@@ -35,7 +35,7 @@ public class CreateAccount{
         this.emailAddress = emailAddress;
     }
 
-    public long getBalance() {
+    public int getBalance() {
         return balance;
     }
